requirements.txt
python-docx==1.1.2
markdown==3.7 